tests/integration_tests/requirements.txt
aiohttp==3.9.3
python-dotenv==1.0.1
termcolor==2.4.0
